+++ CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+
+### New features
+
+### Bug fixes
+
+* Added fix to avoid tarfile path traversal (CVE-2007-4559). (#3)
+
+### Library updates
+
+
 ## [0.0.7] - 2022-09-26
 
 ### Library updates